+++ client/CHANGELOG.md
@@ -1,6 +1,9 @@
 # Change Log
 All notable changes to the "Harbour and xHarbour" extension will be documented in this file.
 
+## 0.8.4
+  - **Server** fixed crash on completition
+
 ## 0.8.3
  - **Server** added completition and goto definition on include
  - **Server** fixed crash on completition on beginning of file